requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
decorator==5.1.0
Django==3.2.9
django-rest-knox==4.1.0
djangorestframework==3.12.4
future==0.18.2
idna==3.3
ipython==7.29.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
